level merge and corrected

+++ bot/src/handlers.go
@@ -47,7 +47,7 @@ func new_message_posted(sess *discordgo.Session, m *discordgo.MessageCreate) {
 	if err != nil { log.Fatal(err) }
 
 	if len(users) == 0 {
-		level_calculated := int64(calcul_level_with_nb_messages(1))
+		level_calculated := calcul_level_with_nb_messages(1)
 
 		new_user := &level{User_ID: user_id, Guild_ID: guild_id, Level: level_calculated}
 		_, err = db.NewInsert().Model(new_user).Ignore().Exec(ctx)
@@ -60,7 +60,7 @@ func new_message_posted(sess *discordgo.Session, m *discordgo.MessageCreate) {
 		user := users[0]
 		user.Nb_Msg += 1
 		
-		level_calculated := int64(calcul_level_with_nb_messages(user.Nb_Msg))
+		level_calculated := calcul_level_with_nb_messages(user.Nb_Msg)
 
 		if level_calculated > user.Level {
 			if int(level_calculated) > int(user.Level) { levels_message(sess, channel_id, &user, int(level_calculated)) }
